Hapus yang tidak digunakan
