+++ database/seeds/RoleSpatieSeeder.php
@@ -31,8 +31,9 @@
 
 use App\Models\User;
 use Illuminate\Database\Seeder;
-use Spatie\Permission\Models\Permission;
+use Illuminate\Support\Facades\DB;
 use Spatie\Permission\Models\Role;
+use Spatie\Permission\Models\Permission;
 
 class RoleSpatieSeeder extends Seeder
 {
